# code/collections/swagger.yaml
swagger: '2.0'
info:
  description: ''
  version: 1.0.0
  title: 'Swagger JsonApi'
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: localhost
schemes:
  - https
  - http
paths:
  /users/:
    get:
      tags:
        - users
      summary: 'User List'
      operationId: listUser
      produces:
        - application/json
      parameters: {  }
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: '/users?page[number]=1&page[size]=100'
                  first:
                    type: string
                    example: '/users?page[number]=1&page[size]=100'
                  last:
                    type: string
                    example: '/users?page[number]=1&page[size]=100'
                  prev:
                    type: string
                    example: 'null'
                  next:
                    type: string
                    example: 'null'
              data:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
                      example: 12
                    type:
                      type: string
                      example: users
                    attributes:
                      $ref: '#/definitions/User'
                    relationships:
                      properties: {  }
    post:
      tags:
        - users
      summary: 'Add User'
      operationId: addUser
      produces:
        - application/json
      parameters:
        -
          in: body
          name: body
          description: addUser
          required: true
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  type:
                    type: string
                    example: users
                  attributes:
                    $ref: '#/definitions/User'
                  relationships:
                    properties: {  }
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: /users
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: users
                  attributes:
                    $ref: '#/definitions/User'
                  relationships:
                    properties: {  }
  '/users/{User_id}':
    patch:
      tags:
        - users
      summary: 'Edit User'
      operationId: editUser
      produces:
        - application/json
      parameters:
        -
          in: body
          name: body
          description: editUser
          required: true
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: users
                  attributes:
                    $ref: '#/definitions/User'
                  relationships:
                    properties: {  }
        -
          name: User_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: /users
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: users
                  attributes:
                    $ref: '#/definitions/User'
                  relationships:
                    properties: {  }
    delete:
      tags:
        - users
      summary: 'Delete User'
      operationId: deleteUser
      produces:
        - application/json
      parameters:
        -
          name: User_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: /users
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: users
                  attributes:
                    $ref: '#/definitions/User'
                  relationships:
                    properties: {  }
    get:
      tags:
        - users
      summary: 'Get User'
      operationId: viewUser
      produces:
        - application/json
      parameters:
        -
          name: User_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: /users
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: users
                  attributes:
                    $ref: '#/definitions/User'
                  relationships:
                    properties: {  }
  /vouchers/:
    get:
      tags:
        - vouchers
      summary: 'Voucher List'
      operationId: listVoucher
      produces:
        - application/json
      parameters: {  }
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: '/vouchers?page[number]=1&page[size]=100'
                  first:
                    type: string
                    example: '/vouchers?page[number]=1&page[size]=100'
                  last:
                    type: string
                    example: '/vouchers?page[number]=1&page[size]=100'
                  prev:
                    type: string
                    example: 'null'
                  next:
                    type: string
                    example: 'null'
              data:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int64
                      example: 12
                    type:
                      type: string
                      example: vouchers
                    attributes:
                      $ref: '#/definitions/Voucher'
                    relationships:
                      properties: {  }
    post:
      tags:
        - vouchers
      summary: 'Add Voucher'
      operationId: addVoucher
      produces:
        - application/json
      parameters:
        -
          in: body
          name: body
          description: addVoucher
          required: true
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  type:
                    type: string
                    example: vouchers
                  attributes:
                    $ref: '#/definitions/Voucher'
                  relationships:
                    properties: {  }
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: /vouchers
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: vouchers
                  attributes:
                    $ref: '#/definitions/Voucher'
                  relationships:
                    properties: {  }
  '/vouchers/{Voucher_id}':
    patch:
      tags:
        - vouchers
      summary: 'Edit Voucher'
      operationId: editVoucher
      produces:
        - application/json
      parameters:
        -
          in: body
          name: body
          description: editVoucher
          required: true
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: vouchers
                  attributes:
                    $ref: '#/definitions/Voucher'
                  relationships:
                    properties: {  }
        -
          name: Voucher_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: /vouchers
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: vouchers
                  attributes:
                    $ref: '#/definitions/Voucher'
                  relationships:
                    properties: {  }
    delete:
      tags:
        - vouchers
      summary: 'Delete Voucher'
      operationId: deleteVoucher
      produces:
        - application/json
      parameters:
        -
          name: Voucher_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: /vouchers
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: vouchers
                  attributes:
                    $ref: '#/definitions/Voucher'
                  relationships:
                    properties: {  }
    get:
      tags:
        - vouchers
      summary: 'Get Voucher'
      operationId: viewVoucher
      produces:
        - application/json
      parameters:
        -
          name: Voucher_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: 'successful operation'
          schema:
            type: object
            properties:
              jsonapi:
                type: object
                properties:
                  version:
                    type: string
                    example: '1.0'
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: /vouchers
              data:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
                  type:
                    type: string
                    example: vouchers
                  attributes:
                    $ref: '#/definitions/Voucher'
                  relationships:
                    properties: {  }
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
  User:
    type: object
    properties: {  }
  Voucher:
    type: object
    properties:
      code:
        type: string
      description:
        type: text
      price:
        type: number
        format: decimal
      quantityInStock:
        type: integer
        format: int32
      vaildFrom:
        type: string
        format: date-time
      validTo:
        type: string
        format: date-time
externalDocs:
  description: 'Find out more about Swagger'
  url: 'http://swagger.io'